refactor: rename ProductsRepository service to repositories and update imports across controllers and services
