+++ src/main/java/ugr/dss/quick_shop/controllers/AdminController.java
@@ -1,11 +1,6 @@
 package ugr.dss.quick_shop.controllers;
 
 import java.io.IOException;
-import java.io.InputStream;
-import java.nio.file.Files;
-import java.nio.file.Path;
-import java.nio.file.Paths;
-import java.nio.file.StandardCopyOption;
 import java.util.Date;
 import java.util.List;
 
@@ -27,7 +22,7 @@ import ugr.dss.quick_shop.models.product.EditProductDto;
 import ugr.dss.quick_shop.models.product.Product;
 import ugr.dss.quick_shop.models.product.ProductDto;
 import ugr.dss.quick_shop.services.DatabaseExportService;
-import ugr.dss.quick_shop.services.ProductsRepository;
+import ugr.dss.quick_shop.repositories.ProductsRepository;
 import ugr.dss.quick_shop.utils.FileUploadUtil;
 
 @Controller
