+++ src/main/resources/application.properties
@@ -18,8 +18,6 @@ server.port=8080
 h2.console.enabled=true
 h2.console.path=/h2-console
 
-# logging.level.org.thymeleaf=DEBUG
-# logging.level.org.springframework.web=DEBUG
 logging.level.org.springframework.security=DEBUG
 
 file.upload-dir=src/main/resources/static/images/
